pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuzz_generator"
version = "0.1.0"
description = "AI Agent-based Fuzz Test Data Modeling Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Fuzz Generator Team"}
]
keywords = ["fuzz", "testing", "ai", "agent", "code-analysis", "joern"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
]

dependencies = [
    # AutoGen official packages (0.4+ new architecture)
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai]>=0.4.0",
    # HTTP client
    "httpx>=0.25.0",
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Configuration
    "pyyaml>=6.0",
    # CLI framework
    "click>=8.0.0",
    # Logging
    "loguru>=0.7.0",
    # Template engine
    "jinja2>=3.0.0",
    # XML processing
    "lxml>=5.0.0",
    # Async support
    "anyio>=4.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
fuzz-generator = "fuzz_generator.cli:cli"

[project.urls]
Homepage = "https://github.com/your-org/fuzz_generator"
Documentation = "https://github.com/your-org/fuzz_generator#readme"
Repository = "https://github.com/your-org/fuzz_generator"
Issues = "https://github.com/your-org/fuzz_generator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fuzz_generator*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["fuzz_generator"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["fuzz_generator"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

